Change how inputs are handled in `useForm`
The change is made so a referece to the inputs are stored internally in
`useForm`. This makes it so if inputs need to change `useForm.setInputs`
needs to be called manually with the new inputs.
On the plus side we make the hook simpler to use as the user no longer
needs to make sure the inputs object they pass keeps it referential
identity.