Add functionality for handling nested inputs object
Now it's possible to pass an `inputs` object in the following ways.

Non-exhaustive list:
- `{ input }`
- `{ list: [ input ] }`
- `{ obj: { input } }`
- `{ list: [ { input } ] }`

It's also possible to set nested object with the `inputs.set` method.
`form.inputs.set("name.nested[0].input", input)`

`form.errors`, `form.changed`, `form.inputErrors`, and `form.values`
will all be the same structure as the inputs. This is also true for the
values passed as arguments to `handleSubmit`.

Submit errors are expected to also have the same shape as the inputs.